Fix mask propagation in the presence of where
Before, when calling ops.where, masks were not properly propagated. We
now restrict the optimisation to `ops.masked`, which I think it was what
the original code intended to do.

I'm not 100% sure that even in the masked case this code is not
introducing some bugs, but this is a strict improvement over the
previous state.

ghstack-source-id: 12466ab827700ae55dc55e3cadc135b1f9dc8ffc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/125574